+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-2pc-spi/src/main/java/io/shardingsphere/transaction/spi/xa/XATransactionManager.java
@@ -17,9 +17,9 @@
 
 package io.shardingsphere.transaction.spi.xa;
 
-import io.shardingsphere.transaction.internal.context.XATransactionContext;
+import io.shardingsphere.transaction.core.internal.context.XATransactionContext;
 import io.shardingsphere.core.rule.DataSourceParameter;
-import io.shardingsphere.transaction.internal.manager.ShardingTransactionManager;
+import io.shardingsphere.transaction.core.internal.manager.ShardingTransactionManager;
 
 import javax.sql.DataSource;
 import javax.sql.XADataSource;
