+++ sharding-transaction/sharding-transaction-2pc/sharding-transaction-xa/src/main/java/io/shardingsphere/transaction/xa/handler/XAShardingTransactionHandler.java
@@ -18,9 +18,9 @@
 package io.shardingsphere.transaction.xa.handler;
 
 import io.shardingsphere.core.constant.transaction.TransactionType;
-import io.shardingsphere.transaction.internal.context.XATransactionContext;
+import io.shardingsphere.transaction.core.internal.context.XATransactionContext;
 import io.shardingsphere.transaction.core.handler.ShardingTransactionHandlerAdapter;
-import io.shardingsphere.transaction.internal.manager.ShardingTransactionManager;
+import io.shardingsphere.transaction.core.internal.manager.ShardingTransactionManager;
 import io.shardingsphere.transaction.xa.manager.XATransactionManagerSPILoader;
 
 /**
