+++ sharding-transaction/sharding-transaction-core/src/main/java/io/shardingsphere/transaction/core/internal/context/ShardingTransactionContext.java
@@ -15,7 +15,7 @@
  * </p>
  */
 
-package io.shardingsphere.transaction.internal.context;
+package io.shardingsphere.transaction.core.internal.context;
 
 import io.shardingsphere.core.constant.transaction.TransactionOperationType;
